requirements.txt
streamlit 
python-dotenv
PyPDF2 
streamlit-extras 
langchain 
spacy 
openai 
faiss-cpu
tiktoken
langchain-community